database implementation and encapsulation
set to create the whole mysql database when you open 'index.php' on the
running wamp server. this has the two table for login and cards and the
php files have been refactored with the correct names. A few cards are
in the database but without unique values so we need to need a few more
at least